% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DegreesFromNorth.r
\name{DegreesFireRun}
\alias{DegreesFireRun}
\title{Get Direction from North}
\usage{
DegreesFireRun(x1, x2, y1, y2)
}
\arguments{
\item{x1}{x coordinate of origin point}

\item{x2}{x coordinate of destination point}

\item{y1}{y coordinate of origin point}

\item{y2}{y coordinate of destination point}
}
\value{
a direction value of a vector calculated from North
}
\description{
Function for calculate the angle of the fire run
#'
}
\examples{
x1<-3
y1<-3
x2<-4
y2<-4
DegreesFireRun(x1,y1,x2,y2)

}
