% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edges.r
\name{Edges}
\alias{Edges}
\title{Create Fronts object}
\usage{
Edges(AllPols, nameFeho)
}
\arguments{
\item{AllPols}{Isochrone map Spatvector polygon object}

\item{nameFeho}{Name of hour ID field in polygon file}
}
\value{
a Spatvector object with lines between polygons (fronts)
}
\description{
Function for create a Vector that returns the front between polygons in an isochrone map
}
