% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveHoles.r
\name{remove.holes.withinFront}
\alias{remove.holes.withinFront}
\title{Remove holes in polygon but keep previous front}
\usage{
remove.holes.withinFront(x, Front)
}
\arguments{
\item{x}{spatVector object corresponding to the evaluated polygon}

\item{Front}{spatVector line corresponding to the previous front}
}
\value{
a spatVector object without holes except for the previous front
}
\description{
Function for make a polygon without holes if the front is all within the polygon, so keeping the previous front 'hole'
#'
}
